requirements.txt
flask
transformers
gtts
pillow
gunicorn
torchvision
torchaudio
torch
gtts
pillow
sentencepiece